Refactor childnutrition to lamp project name and add github workflow file
