+++ api/pom.xml
@@ -3,12 +3,12 @@
 	<modelVersion>4.0.0</modelVersion>
 	<parent>
 		<groupId>org.openmrs.module</groupId>
-		<artifactId>childnutrition</artifactId>
+		<artifactId>lamp</artifactId>
 		<version>1.0.0-SNAPSHOT</version>
 	</parent>
 
-	<artifactId>childnutrition-api</artifactId>
+	<artifactId>lamp-api</artifactId>
 	<packaging>jar</packaging>
-	<name>Child nutrition API</name>
-	<description>API project for Child nutrition</description>
+	<name>Lamp API</name>
+	<description>API project for Lamp</description>
 </project>
