+++ omod/pom.xml
@@ -3,19 +3,19 @@
 	<modelVersion>4.0.0</modelVersion>
 	<parent>
 		<groupId>org.openmrs.module</groupId>
-		<artifactId>childnutrition</artifactId>
+		<artifactId>lamp</artifactId>
 		<version>1.0.0-SNAPSHOT</version>
 	</parent>
 
-	<artifactId>childnutrition-omod</artifactId>
+	<artifactId>lamp-omod</artifactId>
 	<packaging>jar</packaging>
-	<name>Child nutrition OMOD</name>
-	<description>Omod submodule for Child nutrition</description>
+	<name>Lamp OMOD</name>
+	<description>Omod submodule for Lamp</description>
 
 	<dependencies>
         <dependency>
             <groupId>org.openmrs.module</groupId>
-            <artifactId>childnutrition-api</artifactId>
+            <artifactId>lamp-api</artifactId>
             <version>1.0.0-SNAPSHOT</version>
         </dependency>
 		<dependency>
